fixed errors for getting data from api for wine details
