--- HCI_mini_project/src/app/app.component.css
@@ -1,39 +1,0 @@
-div {
-    display: grid; 
-    grid-template-columns: 25% 75%;
-    grid-template-rows: 10% 90%;
-    width: 100%;
-    height: 100%;
-}
-
-app-search {
-    grid-column: 1;
-    grid-row: 1 / span 2;
-}
-
-app-filter {
-    grid-column: 2;
-    grid-row: 1;
-}
-
-app-homepage {
-    grid-column: 2;
-    grid-row: 2;
-    padding: 15px;
-}
-
-
-@media screen and (max-width: 1160px) {
-    div {
-        display: flex;
-        flex-direction: column;
-    }
-    app-search {
-        display: none;
-    }
-
-    app-filter {
-        width: 100%;
-    }
-}
-
