+++ HCI_mini_project/src/app/app.component.html
@@ -1,10 +1,1 @@
-<!-- <app-homepage></app-homepage>
 <router-outlet></router-outlet> 
-<div>
-  <app-filter></app-filter>
-  <app-search></app-search>
-  <app-homepage></app-homepage>
-</div>
-</div> -->
-<!-- <app-homepage></app-homepage> -->
-<!-- <router-outlet></router-outlet>  -->
