+++ HCI_mini_project/src/app/app.module.ts
@@ -5,7 +5,7 @@ import { MaterialModule } from 'src/infrastructure/material.module';
 import { MAT_FORM_FIELD_DEFAULT_OPTIONS, MatFormFieldModule } from '@angular/material/form-field';
 import { AppRoutingModule } from 'src/infrastructure/app-routing.module';
 import { AppComponent } from './app.component';
-import { HomepageComponent } from './homepage/homepage.component';
+import { RecipeCardsComponent } from './recipe-cards/recipe-cards.component';
 import { SearchComponent } from './search/search.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
@@ -16,15 +16,17 @@ import { CommonModule } from '@angular/common';
 import {MatSnackBarModule} from '@angular/material/snack-bar';
 import { FilterComponent } from './filter/filter.component'
 import { HttpClientModule } from '@angular/common/http';
+import { HomepageComponent } from './homepage/homepage.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     DetailsHeaderComponent,
     RecipeDetailsComponent,
-    HomepageComponent,
+    RecipeCardsComponent,
     SearchComponent,
-    FilterComponent
+    FilterComponent,
+    HomepageComponent
   ],
   imports: [
     BrowserModule,
