+++ HCI_mini_project/src/app/filter/filter.component.ts
@@ -105,7 +105,7 @@ export class FilterComponent implements OnInit {
   openFilter(): void {
     const search = document.getElementsByTagName('app-search')[0] as HTMLElement;
     const filter = document.getElementsByTagName('app-filter')[0] as HTMLElement;
-    const homep = document.getElementsByTagName('app-homepage')[0] as HTMLElement;
+    const homep = document.getElementsByTagName('app-recipe-cards')[0] as HTMLElement;
 
     if(filter) {
       search.style.display = "block";
