+++ HCI_mini_project/src/app/recipe-cards/recipe-cards.component.css
@@ -1,0 +1,153 @@
+.recipe-card {
+    height: 240px;
+    width: 250px;
+    cursor: pointer;
+    position: relative;
+    
+}
+
+.recipe-card:hover {
+    opacity: 0.8;
+}
+
+.recipe-card-image {
+    height: 71%;
+    width: 100%;
+    border-radius: 15px 15px 0px 0px;
+    /* object-fit: cover; */
+    background-size: 110% 130%;
+    background-position: center;
+}
+
+.recipe-card-text {
+    /* height: fit-content;
+    font-weight: 700;
+    color: #B9BBBD;
+    padding: 10px;
+    margin-top: -10px;
+    padding-bottom: 15px; */
+    height: fit-content;
+    font-weight: 700;
+    color: #B9BBBD;
+    padding: 10px;
+    margin-top: -50px;
+    /* padding-bottom: 15px; */
+    background-color: white;
+    border-radius: 0px 0px 15px 15px;
+    z-index: 100;
+    width: -webkit-fill-available;
+    /* width: 100%; */
+    position: absolute;
+    bottom: 0;
+}
+
+.recipe-card-title {
+    font-size: 1em;
+}
+
+#card-container {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr 1fr;
+    grid-template-rows: 1fr 1fr;
+    gap: 20px;
+    width: fit-content;
+    padding: 20px;
+    padding: 5px;
+    /* width: 70%; */
+}
+
+#paginated-cards-container {
+    height: 100%;
+    display: grid;
+    grid-template-rows: 90% 10%;
+    width: fit-content;
+    margin: 0 auto;
+}
+
+::ng-deep mat-paginator {
+    color: #898989 !important;
+    cursor: pointer;
+    margin-top: 20px;
+}
+
+::ng-deep .mat-mdc-paginator {
+    background-color: transparent;
+    margin: 0 auto;
+}
+
+.no-results {
+    color: #898989 !important;
+    text-align: left;
+}
+
+@media screen and (max-width: 1450px) {
+    #card-container {
+        grid-template-columns: 1fr 1fr 1fr;
+        grid-template-rows: 1fr 1fr;
+        margin-top: 65px;
+    }
+
+    ::ng-deep .mat-mdc-paginator {
+        margin: 0 auto;
+        margin-top: 17px;
+    }
+
+    .no-results {
+        margin-top: 65px;
+    }
+    
+}
+
+@media screen and (max-width: 1160px) {
+    #card-container {
+        grid-template-columns: 1fr 1fr 1fr 1fr;
+        grid-template-rows: 1fr 1fr;
+        margin-top: 0px;
+    }
+
+    ::ng-deep .mat-mdc-paginator {
+        margin: 0 auto;
+    }
+
+    .no-results {
+        margin-top: 0px;
+    }
+    
+}
+
+@media screen and (max-width: 1100px) {
+    #card-container {
+        grid-template-columns: 1fr 1fr 1fr;
+        grid-template-rows: 1fr 1fr;
+        margin-top: 0px;
+    }
+    
+}
+
+@media screen and (max-width: 840px) {
+    #card-container {
+        grid-template-columns: 1fr 1fr;
+        grid-template-rows: 1fr 1fr 1fr;
+    }
+
+    #paginated-cards-container {
+        grid-template-rows:  95% 5%;
+    }
+}
+
+@media screen and (max-width: 565px) {
+    #card-container {
+        grid-template-columns: 1fr;
+        grid-template-rows: 1fr 1fr 1fr 1fr 1fr;
+        
+    }
+
+    #paginated-cards-container {
+        grid-template-rows:  98% 2%;
+    }
+
+    .recipe-card {
+        width: 300px;
+    }
+}
+
